Technical/warnings2

+++ lib/ZipFile/ZipReader.h
@@ -50,7 +50,7 @@ public:
 
 	bool seek(int offset, SeekOrigin origin = SeekOrigin::Start);
 
-	size_t read(int8_t *d, uint32_t sz);
+	size_t read(int8_t *d, size_t sz);
 	std::vector<int8_t> read();
 	void extract(const std::string &path);
